UI: Separate delegate class into header file
